move controller to different namespace
